+++ app/src/main/java/com/purwasadr/pantaucovid/data/source/remote/response/TotalResponse.kt
@@ -4,14 +4,14 @@ import com.google.gson.annotations.SerializedName
 
 data class TotalResponse(
     @field:SerializedName("jumlah_meninggal")
-    val jumlahMeninggal: Int? = null,
+    val jumlahMeninggal: Long? = null,
 
     @field:SerializedName("jumlah_sembuh")
-    val jumlahSembuh: Int? = null,
+    val jumlahSembuh: Long? = null,
 
     @field:SerializedName("jumlah_positif")
-    val jumlahPositif: Int? = null,
+    val jumlahPositif: Long? = null,
 
     @field:SerializedName("jumlah_dirawat")
-    val jumlahDirawat: Int? = null
+    val jumlahDirawat: Long? = null
 )
